`./ts` file restructure
