+++ src/ts/events/index.ts
@@ -1,3 +1,6 @@
+// Re-export facade
+export { Events } from './facade';
+
 // Re-export all event handler modules
 export { DragHandler } from './drag';
 export { PinchHandler } from './pinch';
@@ -6,7 +9,7 @@ export { GestureHandler } from './gesture';
 export { WheelHandler } from './wheel';
 export { KeyboardHandler } from './keyboard';
 export { DoubleTapHandler } from './doubletap';
-export { UpdateHandler, UpdateEvents } from './update';
+export { UpdateHandler } from './update';
 
 // Re-export shared types and utilities
 export {
