+++ src/ts/i18n/index.ts
@@ -1,0 +1,2 @@
+export { i18n, langs } from './strings';
+export { locale, languageNames, rtlLanguageCodes } from './locale';
