+++ src/types/models.ts
@@ -2,9 +2,9 @@
 /// <reference types="svelte" />
 
 import type { HTMLMicrioElement } from '../ts/element';
-import type { Grid } from '../ts/grid';
+import type { Grid } from '../ts/nav/grid';
 import type { MicrioImage } from '../ts/image';
-import type { VideoTourInstance } from '../ts/videotour';
+import type { VideoTourInstance } from '../ts/media/videotour';
 import type { Writable } from 'svelte/store';
 import type { MicrioIcon } from '../ts/icons';
